refactor(output_schema): decompose metadata module
Split the ~3.5k-line src/output_schema/metadata.rs documentation registry
into a metadata/ directory module with cohesive submodules:

- document: top-level output, header, extra-data, system-environment,
  license-index-provenance field docs
- summary: summary, tallies, facet-tallies, tally-entry, clarity-score
- file_info: per-file/directory output surface
- file_evidence: authors, copyrights, emails, holders, urls
- package: package, package-data, party, file-reference
- dependency: dependency, top-level dependency, resolved-package
- license: license-detection, match, license/rule references, policy entry

mod.rs retains the public OutputTypeDoc / OutputFieldDoc structs, the
DOCUMENTED_TYPES registry, documented_output_types(), and the registry
tests. Field-doc consts are pub(super) and re-aggregated in mod.rs, so the
public API (crate::output_schema::metadata::*) is unchanged.

Pure refactor: field-doc and type-doc counts are identical (375 / 33), and
regenerating docs/OUTPUT_FIELD_REFERENCE.md produces no diff. No .unwrap()
calls existed to remediate.
